splitted to services

+++ Secretary.Cache/ICacheService.cs
@@ -2,8 +2,8 @@
 
 public interface ICacheService
 {
-    Task SaveEntity<T>(long key, T value, short? lifetimeSec = 600) where T : class;
-    Task SaveEntity<T>(string key, T value, short? lifetimeSec = 600) where T : class;
+    Task SaveEntity<T>(long key, T value, int? lifetimeSec = 600) where T : class;
+    Task SaveEntity<T>(string key, T value, int? lifetimeSec = 600) where T : class;
     Task<T?> GetEntity<T>(long key) where T : class;
     Task<T?> GetEntity<T>(string key) where T : class;
     Task DeleteEntity<T>(long key) where T : class;
